+++ src/earthen/clj_cb/bucket.clj
@@ -15,7 +15,8 @@
 (def ^{:doc "Wrapper of clojure.data.json/json-str."}
   write-json json/json-str)
 
-(defn content
+(defn content->map
+  "Returns the content of a JsonDocument as a map"
   [json-document]
   (if json-document
     (-> json-document
@@ -23,24 +24,26 @@
         .toString
         read-json)))
 
-(defn document
+(defn document->map
+  "Converts a JsonDocument to a map"
   [jsondocument]
   {:id (.id jsondocument)
    :cas (.cas jsondocument)
    :expiry (.expiry jsondocument)
    :mutation-token (.mutationToken jsondocument)
-   :content (content jsondocument)})
+   :content (content->map jsondocument)})
 
 
-(defn create-counter
-  "Increment or decrement a counter with 0 value
-  and a default value of 20 with the default key/value timeout."
+(defn create-counter!
+  "Increment or decrement a counter. If only key given, will take initial value as 0 (if not exists) and increment with 1"
   ([bucket k ]
    (create-counter bucket k 0 1))
   ([bucket k delta initial]
    (.counter bucket k delta initial)))
 
 (defn get
+  "Retrieves a document from the bucket. By default :json map document is returned.
+  :raw is the string json"
   ([bucket id]
    (get bucket id :json))
   ([bucket id format]
@@ -50,20 +53,21 @@
          (-> doc
              .content
             .toString)
-         (document doc))))))
+         (document->map doc))))))
 
 (defn get-and-lock
+  "Retrieves and locks the document for n seconds"
   [bucket id seconds]
   (let [doc (.getAndLock bucket id seconds)]
     (if doc
-       (document doc))))
+       (document->map doc))))
 
-(defn touch
+(defn touch!
   "Renews the expiration time of a document with the default key/value timeout"
   [bucket id expiry]
   (.touch bucket id expiry))
 
-(defn get-and-touch
+(defn get-and-touch!
   "Retrieve and touch a JsonDocument by its unique ID with the default key/value timeout."
   ([bucket id] (get-and-touch bucket id 0))
   ([bucket id expiry]
@@ -74,25 +78,30 @@
   (.unlock bucket (:id document) (:cas document)))
 
 (defn create-json-document
-  [id json]
-  (let [json-object (JsonObject/fromJson (write-json json))]
+  "Creates a JsonDocument from a json-map"
+  [id json-map]
+  (let [json-object (JsonObject/fromJson (write-json json-map))]
     (JsonDocument/create id json-object)))
 
-(defn replace
-  [bucket id json]
-  (let [json (if (string? json) (read-json json) json)
+(defn replace!
+  "Save or delete a document from the bucket"
+  [bucket id json-map]
+  (let [json (if (string? json-map) (read-json json-map) json-map)
         doc (create-json-document id json)]
-    (document (.upsert bucket doc))))
+    (document->map (.upsert bucket doc))))
 
 (defn manager
+  "Returns the manager giving a bucket"
   [bucket]
   (.bucketManager bucket))
 
-(defn remove
+(defn remove!
+  "Remove the id from the bucket"
   [bucket id]
   (.remove bucket id))
 
 (defn close
+  "Closes the bucket connection"
   ([bucket] (close bucket 30 :SECONDS))
   ([bucket time type]
    (.close bucket time (u/time type))))